requirements.txt
streamlit
pymupdf
pandas
numpy
opencv-python-headless
Pillow
openai
pydantic
openpyxl
matplotlib
scikit-image